UseThy operation should observe results
Previously, `Operation.UseThy` would consume markup, but strangely did
not observe the final result. This means that errors in theories may not
propagate properly. This commit makes UseThy fail accordingly.
